db: WIP: Redo statement and match synchronization & abort
Mostly works again but crashes on collision. Two big changes: 1. we
null out the child lists in matches and statements before we clear
them, so that when you insert a child, you're guaranteed that it will
actually get cleared. 2. we reorganize insertion logic so any failure
aborts the whole match/statement creation (?)

+++ db.h
@@ -103,8 +103,8 @@ StatementRef dbInsertOrReuseStatement(Db* db, Clause* clause,
 // so that the thread can be interrupted if the match is
 // destroyed. The Match is returned acquired and needs to be released
 // by the caller.
-Match* dbInsertMatch(Db* db, size_t nParents, StatementRef parents[],
-                       pthread_t workerThread);
+Match* dbInsertMatch(Db* db, int nParents, StatementRef parents[],
+                     pthread_t workerThread);
 
 void dbRetractStatements(Db* db, Clause* pattern);
